input/elasticapm: reset json-iterator correctly (#8)
When resetting a NDJSONStreamDecoder, make sure
we don't read buffered data from a reused json-
iterator decoder. This could happen if there is
some junk data after an event object, but before
the newline delimiter.

We fix this by not using json-iterator's Decoder
type, but instead borrowing and releasing an
Iterator with the buffered line.