test: fix data race in TestConcurrentAsync
We were previously using a struct so this wasn't obvious. Now that we
are passing a pointer we can't share the modelpb.APMEvent because the
readMetadata is called before cloning the base event.
The data race only happens in the test because of how it's written.
Production code (APM Server) is not sharing APMEvents.