Remove APMEvent.Processor, infer type
Remove the APMEvent.Processor type, which is a relic
of the past when APM Server had different event processors
and we set this field according to which processor created
the document.

Instead of setting these fields explicitly, we can infer
the type of an event from other fields:
 - if Metricset is non-nil, it's a metric
 - if Error is non-nil, it's an error
 - if Log is non-nil, it's a log
 - if Transaction.Type is set, it's a transaction
 - if Span.Type is set, it's a span

One additional rule, since there may be logs without `log.*`
fields (hence Log will be nil): if `event.kind` is "event",
then we treat that as a log as well.

We still use `processor.*` fields in various places in
the UI. Until/unless we remove them from there as well,
we can infer them during ingest, or set them as
constant_keyword values in the mapping based on the datastream.