properly test Received
`Event.Received` weren't really tested by the current tests, due to the
use of `cmpopts.IgnoreUnexported`. I wrongly assumed that with that
option only unexported fields would be ignored but in this case the
result was a complete lack of verification on the time value.

This commimt removes `IgnoreUnexported` and add a custom comparer to
ensure time correspond.

Co-authored-by: kruskall <99559985+kruskall@users.noreply.github.com>