perf: improve batch pool memory usage
Share the pool among processor for more efficient usage.
Clear the slice before putting it back into the pool.

Because the entire backing array is kept in memory even when
slicing, the modelpb.Batch is keeping around references to
pointers, preventing APMEvents from being garbage collected.
This wasn't a big issue before we switched to protobuf models
because the model.Batch was not a slice of pointers.