Setting event timestamp from OTel observed timestamp when needed (#166)
When receiving an OTel log, sometimes the timestamp might not be available as explained here. Currently, in those cases, the logs are recorded with the timestamp set to the unix epoch start time, which has already caused issues to our customers since they can't find their logs in Kibana within the right time range.

These changes aim to check when the timestamp is not set so that the "observed timestamp" is used instead.