otlp: follow semconv for exception fields
otel semconv require that at lease one field between
exception.message or exception.type to be present on an application
exception.

To ensure we respect them we default on adding exception.type only
if has a value and defaulting exception.message to a constant
string when empty and no exception.type is present.