Require only one of exception.{type,message} for log exceptions (#182)
* otlp: follow semconv for exception fields

otel semconv require that at lease one field between
exception.message or exception.type to be present on an application
exception.

To ensure we respect them we default on adding exception.type only
if has a value and defaulting exception.message to a constant
string when empty and no exception.type is present.

* otlp: add exception when message or type is present

Trigger exception creation for any log message with either message or
type. This aligns to otel semconv requirements for exceptions.