perf: do not pool on each iteration when processing the same batch
The processor is reading batchSize items at a time but on each
iteration it retrieves and submit a batch to the pool. Now that we
don't have the async code handling anymore we can just reuse the batch
and only submit it to the pool once we're done reading the events.