Ensure more unique grouping key when parsed stack trace is not available.
OpenTelemetry today just sends a tree of `exception.message` and `exception.type`

The current logic for `haveExceptionStackTrace` determines there is a stack trace if there is an
exception tree.

This PR ensures that when an exception tree has no stacktrace anywhere in the tree we include
`exception.message` into the hash so that we don't just hash over `exception.type`.